requirements.txt
biopython
polars
lancedb
umap-learn
hdbscan
json-rpc
numpy
scikit-learn
pyarrow
torch
transformers
einops
tenacity
tqdm
PySide6
pyvista
pyvistaqt
